Create cleveland_all_july.py
I think this is my best one yet.  The answer I got from running this code is that it will cost about 256 billion dollars on average for a 100% renewable grid to get Cleveland through July.  If we spent 300 or 350 billion dollars I guess it would be a more reliable grid.